requirements.txt
sqlalchemy~=1.4.40
sqlalchemy-utils
python-dotenv~=0.20.0
psycopg2-binary
pytest
requests~=2.28.1